Revision of way of selecting subregion within region
Because the previous method required the creation of an unnecessary feature class for the subregion, it now creates a feature layer instead (that is temporary and will vanish at the end of the session).

Also got rid of the use of the index of the parameters to retrieve them from the list, and prefers an assignment as a header of the execute method.