Defining relative paths + Defining default outputs
In order to be able to write files in the output GDB, decided to used a default directory structure with default GDBs, that the user can change nonetheless.

Also it does not ask the user for the name of the output features classes, they become fixed, but the user can now choose a prefix for all output feature classes using the 'project name' parameter.