Rolling back to Select instead of SelectByAttribute
While SelectLayerByAttribute_management works perfectly and avoids the creation of a feature class, it is often very slow to run a script on the selected feature layer it generates. For some reason, it seems to run the geoprocessing on the whole dataset, eventhough it only outputs loads for the relevant selected area.

So decided to roll back to the use of Select_analysis that creates a feature class in the output GDB. For now, the feature class is not deleted after geoprocessing, but it could be done in the future.