avoids issue with NULL fields
if a sub-model does not have any feature in a given polygon, it will yield a NULL value in the corresponding attribute, which was a problem in the calculation of the statistics in the postprocessing tool, because 1 + NULL  = NULL, when you want it to = 1. Now fixed by checking if each field is 'None' (i.e. NULL in Python) and replacing it by 0 if it is.