reorders dataframe for loads to keep predefined order
Using the stack method automatically sorts the columns that are stacked as indices, so needed to reorder given the order of the 'load_headers' attribute.

Also there was a mistake where source_fields (which is already a list) was converted to a list. This was only relevant for the df_area.