replaces pickle/unpickle by read/write CSV
Because pickle/unpickle seemed very sensitive to package versions and encoding, replaced the pickle/unpickle by a read/write CSV functionality.

It has the added benefit of generating human-readable files. One of the drawbacks is issues with floating point numbers which may affect some rounding for the 8th decimal number (so  it is a marginal impact).