removes fields for WWTP tool v2 from hard coded, and brings them as parameters
The fieldnames for the treated outflow and the storm water overflow from wastewater treatment plants (WWTPs) were hard-coded in the geo tool. This meant that a dataset using different fieldnames, or more importantly, a dataset gathering several years of data could not be used.

Now, similar to what is done for the Corine Land Cover field, it becomes a parameter for ScenarioV3.run(), and also it adds additional fields with default values for the ArcMap side of SLAMpy.