removes field dependency in WWTP v2 and v3
The dependency of the field on a feature class meant that the "trick" of asking the user to put {} where N or P should be inserted would not work on the ArcMap side, because it would check for the existence of the field in the feature class it is depending on. To remove this issue, the dependency is removed.